week 7 assignment 1 backendconverted to ts successfully
