+++ week-7/assignment-1/db/index.ts
@@ -1,0 +1,17 @@
+// const mongoose = require("mongoose");
+import mongoose from "mongoose";
+
+const userSchema = new mongoose.Schema({
+  username: String,
+  password: String,
+});
+
+const todoSchema = new mongoose.Schema({
+  title: String,
+  description: String,
+  done: Boolean,
+  userId: String,
+});
+
+export const User = mongoose.model("User", userSchema);
+export const Todo = mongoose.model("Todo", todoSchema);
