--- week-7/assignment-1/middleware/index.js
@@ -1,24 +1,0 @@
-const jwt = require('jsonwebtoken');
-const { Response } = require('express');
-const SECRET = 'SECr3t';  // This should be in an environment variable in a real application
-
-const authenticateJwt = (req, res, next) => {
-  const authHeader = req.headers.authorization;
-  if (authHeader) {
-    const token = authHeader.split(' ')[1];
-    jwt.verify(token, SECRET, (err, user) => {
-      if (err) {
-        return res.sendStatus(403);
-      }
-      req.userId = user.id;
-      next();
-    });
-  } else {
-    res.sendStatus(401);
-  }
-};
-
-module.exports = {
-    authenticateJwt,
-    SECRET
-}
